requirements.txt
langchain
docling
streamlit
faiss-cpu
langchain-community
langchain-openai
python-dotenv
youtube-transcript-api
beautifulsoup4